requirements1.txt
# Updated requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langchain-groq>=0.1.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
pdf2image>=1.16.3
pillow>=10.1.0
pytesseract>=0.3.10
python-docx>=0.8.11
tqdm>=4.66.1
ollama>=0.1.0
requests>=2.31.0
faiss-cpu>=1.7.4
python-magic>=0.4.27
filetype>=1.2.0
streamlit-chat>=0.1.0
google-generativeai==0.8.3